+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## UNRELEASED
 
 - [README](README.md): Avoid shadowing a variable in the example.
+- `character`: Normalize the server ID to lower case.
 
 ## [v0.0.6](https://github.com/ion1/swtor-settings-updater/releases/tag/v0.0.6) – 2021-12-28